requirements.txt
fastapi
uvicorn
cryptography
pyjwt
